PostCard: Prognostic covariate adjustment for estimation of estimands

Convenient functions for conducting analysis using prognostic covariate adjustment for GLMs. The package uses plug-in estimation for robust estimation of the estimand and influence functions for robust estimation of the variance of the estimand.

See also

Author

Maintainer: Mathias Lerbech Jeppesen mathiasljeppesen@outlook.com

Authors:

Emilie Hoejbjerre-Frandsen ehfd@novonordisk.com