Generate data simulated from a GLM

Provide a formula, variables and a family to generate a linear predictor using the formula and provided variables before using the inverse link of the family to generate the GLM modelled mean, mu, which is then used to simulate the response with this mean from the generating function according to the chosen family.

Usage

glm_data(
  formula_eta,
  ...,
  family = gaussian(),
  family_args = list(sd = 1),
  response_name = "Y"
)

Arguments

formula_eta

an expression specifying how to generate the mean of the response (together with the inverse link from family)

...

a data.frame with columns corresponding to variables used in formula_eta, a named list of those variables, or individually provided named arguments of variables from

family

the family of the response. this can be a character string naming a family function, a family function or the result of a call to a family function

family_args

a named list with values of arguments passed to family relevant r<family_name> function for simulating the data

response_name

a character giving the name of the simulated response

Value

a data.frame

Examples

# Generate a gaussian response from a single covariate
glm_data(1+2*x1,
                x1 = rnorm(10))
#>             Y         x1
#> 1  -2.0993523 -0.3597136
#> 2   1.3803117 -0.1086270
#> 3   4.7834135  1.8277865
#> 4  -4.6407073 -1.9619177
#> 5   2.9280752  0.9459828
#> 6   3.1051919  0.5489621
#> 7   0.5581786 -0.5774653
#> 8  -2.5665320 -1.2282849
#> 9   3.2284699  1.0395321
#> 10  1.5250613  0.4556020

# Generate a gaussian response from a single covariate with
# non-linear effects
glm_data(1+2*abs(sin(x1)),
                x1 = runif(10, min = -2, max = 2))
#>           Y         x1
#> 1  2.982812 -0.3645845
#> 2  2.765510 -0.5565052
#> 3  3.192723  0.6100402
#> 4  1.768523 -1.4257631
#> 5  2.703715  0.7231945
#> 6  2.914777  1.4371436
#> 7  3.233507  1.3545200
#> 8  3.328234  1.5252517
#> 9  3.018810 -1.0492090
#> 10 2.047321  0.3145234

# Generate a negative binomial response
glm_data(1+2*x1-x2,
                x1 = rnorm(10),
                x2 = rgamma(10, shape = 2),
                family = MASS::negative.binomial(2))
#>    Y         x1         x2
#> 1  1 -0.1209674 0.37081227
#> 2  1  0.4582423 2.04115261
#> 3  0  0.6165347 3.69306068
#> 4  0 -0.1750488 1.97369911
#> 5  0 -0.1228313 2.40876607
#> 6  0 -1.4726915 3.06537971
#> 7  0  0.1236927 0.05520964
#> 8  0 -1.0955924 1.82646237
#> 9  1  0.3883423 3.71799172
#> 10 0 -0.7550303 1.30834183

# Provide variables as a list/data.frame
glm_data(1+2*x1-x2,
                data.frame(
                  x1 = rnorm(10),
                  x2 = rgamma(10, shape = 2)
                ),
                family = MASS::negative.binomial(2))
#>    Y          x1         x2
#> 1  0 -0.15401329 0.09544077
#> 2  0 -0.05892959 0.65405248
#> 3  0 -0.53258072 0.26243470
#> 4  2  1.07907227 1.72699254
#> 5  2  0.87461531 0.76655304
#> 6  0 -0.66601844 1.33477946
#> 7  0 -1.12627593 1.11398868
#> 8  0 -0.36943564 0.81570337
#> 9  0 -1.18159506 0.97868209
#> 10 1  0.05931762 2.24684486