Guenther-Schouten Power Approximation

This function calculates the Guenther-Schouten power approximation for ANOVA or ANCOVA. The approximation is based in (Guenther WC. Sample Size Formulas for Normal Theory T Tests. The American Statistician. 1981;35(4):243–244) and (Schouten HJA. Sample size formula with a continuous outcome for unequal group sizes and unequal variances. Statistics in Medicine. 1999;18(1):87–91).

Usage

power_gs(
  n = 100,
  r = 1,
  sigma = sqrt(2),
  rho = NULL,
  R2 = NULL,
  ate = 0.6,
  margin = 0,
  alpha = 0.05
)

Arguments

n

Number of participants in total. From this number of participants in the treatment group is \(n1=(r/(1+r))n\) and the control group is \(n1=(1/(1+r))n\).

r

Allocation ratio \(r=n1/n0\). For one-to-one randomisation r=1.

sigma

Standard deviation of \(Y(w)\), where w is the treatment indicator, and we assume homoskedasticity.

rho

Correlation between outcome and adjustment covariate in univariable covariate adjustment.

R2

The estimated pooled multiple correlation coefficient between outcome and covariates.

ate

Minimum effect size that we should be able to detect.

margin

Superiority margin (for non-inferiority margin, a negative value can be provided).

alpha

Significance level. Due to regulatory guidelines when using a one-sided test, half the specified significance level is used. Thus, for standard alpha = .05, a significance level of 0.025 is used.

Value

a numeric with power approximation based on the Guenther-Schouten approximation

Details

The formula in the case of an ANCOVA model with multiple covariate adjustement is:

$$ n=\frac{(1+r)^2}{r}\frac{(z_{1-\alpha/2}+z_{1-\beta})^2\sigma^2(1-R^2)}{(\beta_1-\beta_0-\Delta_s)^2}+\frac{(z_{1-\alpha/2})^2}{2} $$

where \(\widehat{R}^2\coloneqq \frac{\widehat{\sigma}_{XY}^\top \widehat{\Sigma}_X^{-1}\widehat{\sigma}_{XY}}{\widehat{\sigma}^2}\), we denote by \(\sigma^2\) the variance of the outcome, \(\Sigma_X\) the covariance matrix of the covariates, and \(\sigma_{XY}\) a \(p\)-dimensional column vector consisting of the covariance between the outcome variable and each covariate. In the univariate case \(R^2\) is replaced by \(\rho^2\)

As a default, both R2 and rho are NULL, meaning that power is approximated in the case of an ANOVA with no adjustment covariates are present besides the binary group indicator. In this case, the formula above applies but where \(R^2\) is replaced with \(0\).

Examples

# Approximate the power for an ANCOVA with a single adjustment covariate
power_gs(rho = 0.7)
#> [1] 0.8369041

#' # Approximate power for an ANCOVA with several adjustment covariates
power_gs(R2 = 0.8)
#> [1] 0.9969063

# Approximate the power for an ANOVA with 2:1 randomisation, an assumed
# variance of Y(w) of 4, an assumed effect size of 1.5 and a margin of 1
power_gs(n = 400, r = 2/3, sigma = sqrt(4), ate = 1.5, margin = 1)
#> [1] 0.6856785