Power Approximation based on non-centrality parameter

This function calculates the power for ANOVA or ANCOVA based on the non-centrality parameter and the exact t-distributions.

Usage

power_nc(
  n = 100,
  r = 1,
  sigma = sqrt(2),
  rho = NULL,
  R2 = NULL,
  n.adj = ifelse(is.null(rho), 0, 1),
  ate = 0.6,
  margin = 0,
  alpha = 0.05
)

Arguments

n

Number of participants in total. From this number of participants in the treatment group is \(n1=(r/(1+r))n\) and the control group is \(n1=(1/(1+r))n\).

r

Allocation ratio \(r=n1/n0\). For one-to-one randomisation r=1.

sigma

Standard deviation of \(Y(w)\), where w is the treatment indicator, and we assume homoskedasticity.

rho

Correlation between outcome and adjustment covariate in univariable covariate adjustment.

R2

The estimated pooled multiple correlation coefficient between outcome and covariates.

n.adj

Number of adjustment covariates. Used for calculating the degrees of freedom. Specification only necessary when specifying an R2.

ate

Minimum effect size that we should be able to detect.

margin

Superiority margin (for non-inferiority margin, a negative value can be provided).

alpha

Significance level. Due to regulatory guidelines when using a one-sided test, half the specified significance level is used. Thus, for standard alpha = .05, a significance level of 0.025 is used.

Value

The function returns a power approximation based on the non-centrality parameter and the exact t-distribution.

Details

The prospective power estimations are based on (Kieser M. Methods and Applications of Sample Size Calculation and Recalculation in Clinical Trials. Springer; 2020). The ANOVA power is calculated based on the non-centrality parameter given as

$$nc =\sqrt{\frac{r}{(1+r)^2}\cdot n}\cdot\frac{ate-margin}{\sigma},$$

where we denote by \(\sigma^2\) the variance of the outcome, such that the power can be estimated as

$$1-\beta = 1 - F_{t,n-2,nc}\left(F_{t, n-2, 0}^{-1}(1-\alpha/2)\right).$$

The power of ANCOVA with univariate covariate adjustment and no interaction is calculated based on the non-centrality parameter given as

$$nc =\sqrt{\frac{rn}{(1+r)^2}}\frac{ate-margin}{\sigma\sqrt{1-\rho^2}},$$

such that the power can be estimated as

$$1-\beta = 1 - F_{t,n-3,nc}\left(F_{t, n-3, 0}^{-1}(1-\alpha/2)\right).$$

The power of ANCOVA with either univariate covariate adjustment and interaction or multiple covariate adjustement with or without interaction is calculated based on the non-centrality parameter given as

$$nc =\frac{ate-margin}{\sqrt{\left(\frac{1}{n_1}+\frac{1}{n_0} + X_d^\top\left((n-2)\Sigma_X\right)^{-1}X_d \right)\sigma^2\left(1-\frac{\sigma_{XY}^\top \Sigma_X^{-1}\sigma_{XY}}{\sigma^2}\right)}}.$$

where \(X_d \coloneqq \left(\overline{X}_1^1-\overline{X}_0^1, \ldots, \overline{X}_1^p-\overline{X}_0^p\right)^\top\), \(\widehat{R}^2\coloneqq \frac{\widehat{\sigma}_{XY}^\top \widehat{\Sigma}_X^{-1}\widehat{\sigma}_{XY}}{\widehat{\sigma}^2}\), \(\Sigma_X\) the covariance matrix of the covariates, and \(\sigma_{XY}\) a \(p\)-dimensional column vector consisting of the covariance between the outcome variable and each covariate. Since we are in the case of randomized trials the expected difference between the covariate values between the to groups is 0. Furthermore, the elements of \(\Sigma_X^{-1}\) will be small, unless the variances are close to 0, or the covariates exhibit strong linear dependencies, so that the correlations are close to 1. These scenarios are excluded since they could lead to potentially serious problems regarding inference either way. These arguments are used by Zimmermann et. al (Zimmermann G, Kieser M, Bathke AC. Sample Size Calculation and Blinded Recalculation for Analysis of Covariance Models with Multiple Random Covariates. Journal of Biopharmaceutical Statistics. 2020;30(1):143–159.) to approximate the non-centrality parameter as in the univariate case where \(\rho^2\) is replaced by \(R^2\).

Then the power for ANCOVA with n.adj adjustment covariates can be estimated as

$$1-\beta = 1 - F_{t,n - 2 - n.adj,nc}\left(F_{t, n - 2 - n.adj,0), 0}^{-1}(1-\alpha/2)\right).$$

Examples

# Approximate the power for an ANCOVA with a single adjustment covariate
power_nc(rho = 0.7)
#> [1] 0.8367532

#' # Approximate power for an ANCOVA with several adjustment covariates
power_nc(R2 = 0.8, n.adj = 3)
#> [1] 0.9968886

# Approximate the power for an ANOVA with 2:1 randomisation, an assumed
# variance of Y(w) of 4, an assumed effect size of 3 and a margin of 1
power_nc(n = 400, r = 2/3, sigma = sqrt(4), ate = 1.5, margin = 1)
#> [1] 0.6856692