Fit GLM and find any estimand (marginal effect) using plug-in estimation with variance estimation using influence functions

The procedure uses plug-in-estimation and influence functions to perform robust inference of any specified estimand in the setting of a randomised clinical trial, even in the case of heterogeneous effect of covariates in randomisation groups.

Usage

rctglm(
  formula,
  group_indicator,
  family,
  data,
  group_allocation_prob = NULL,
  estimand_fun = "ate",
  estimand_fun_deriv0 = NULL,
  estimand_fun_deriv1 = NULL,
  verbose = options::opt("verbose"),
  ...
)

Arguments

formula

an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.

group_indicator

(name of) the variable in data that identifies randomisation groups

family

a description of the error distribution and link function to be used in the model. For glm this can be a character string naming a family function, a family function or the result of a call to a family function. For glm.fit only the third option is supported. (See family for details of family functions.)

data

an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which glm is called.

group_allocation_prob

a numeric with the probabiliy of being assigned "group 1" (rather than group 0). As a default, the ratio of 1's in data is used.

estimand_fun

a character specifying a default estimand function, or function with arguments psi0 and psi1 specifying the estimand.

estimand_fun_deriv0

a function specifying the derivative of estimand_fun wrt. psi0. As a default the algorithm will use symbolic differentiation to automatically find the derivative from estimand_fun

estimand_fun_deriv1

a function specifying the derivative of estimand_fun wrt. psi1. As a default the algorithm will use symbolic differentiation to automatically find the derivative from estimand_fun

verbose

numeric verbosity level specifying how much information should be printed to the user (Defaults to 2, overwritable using option 'PostCard.verbose' or environment variable 'R_POSTCARD_VERBOSE')

...

Additional arguments passed to stats::glm()

Value

an rctglm object

Details

The procedure assumes the setup of a randomised clinical trial with observations grouped by a binary group_indicator variable, allocated randomly with probability group_allocation_prob. A GLM is fit and then used to predict the response of all observations in the event that the group_indicator is 0 and 1, respectively. Taking means of these predictions prodeuce the counterfactual means psi0 and psi1, and an estimand r(psi0, psi1) is calculated using any specified estimand_fun.

The variance of the estimand is found by taking the variance of the influence function of the estimand.

This method of inference using plug-in estimation and influence functions for the variance produces a causal estimate of the estimand, as stated by articles XXXX.

Estimands

As noted in the description, psi0 and psi1 are the counterfactual means found by prediction using a fitted GLM in the binary groups defined by group_indicator.

Default estimand functions can be specified via "ate" (taking the difference psi1 - psi0) and "rate_ratio" (taking the ratio psi1 / psi0)

As a default, the Deriv package is used to perform symbolic differentiation to find the derivatives of the estimand_fun.

Examples

# Generate some data to showcase example
n <- 100
dat_binom <- glm_data(
  1+1.5*X1+2*A,
  X1 = rnorm(n),
  A = rbinom(n, 1, .5),
  family = binomial()
)

# Fit the model
ate <- rctglm(formula = Y ~ .,
              group_indicator = A,
              data = dat_binom,
              family = binomial,
              estimand_fun = "ate")
#> ℹ Setting the group allocation probability `group_allocation_prob` as the mean of column `A` in data: 0.48
#> 
#> ── Symbolic differentiation of estimand function ──
#> 
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi0' as: '-1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv0`
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi1' as: '1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv1`