Methods for objects of class rctglm

Methods mostly to extract information from model fit and inference. See details for more information on each method.

Usage

# S3 method for class 'rctglm'
print(x, digits = 3, ...)

# S3 method for class 'rctglm'
coef(object, ...)

# S3 method for class 'rctglm'
summary(object, ...)

estimand(x)

# S3 method for class 'rctglm'
estimand(x)

se_estimand(x)

# S3 method for class 'rctglm'
se_estimand(x)

Arguments

x

an object of class rctglm

digits

a numeric with the number of digits to display when printing

...

additional arguments passed to methods

object

an object of class rctglm

Details

coef and summary just use the corresponding glm methods on the glm fit contained within the rctglm object. Thus, these functions are shortcuts of running coef(x$glm) and summary(x$glm).

estimand and se_estimand are methods for extracting the estimated estimand value as well as the standard error (SE) of the estimand. Extract the plug-in estimation of the estimand, found by using the estimand function on the predicted counterfactual means of each group. These functions are a shortcut to extract the list elements estimand and se_estimandof an rctglm class object.

Examples

# Generate some data to showcase example
n <- 100
dat_binom <- glm_data(
  1+1.5*X1+2*A,
  X1 = rnorm(n),
  A = rbinom(n, 1, .5),
  family = binomial()
)

# Fit the model
ate <- rctglm(formula = Y ~ .,
              group_indicator = A,
              data = dat_binom,
              family = binomial,
              estimand_fun = "ate")
#> ℹ Setting the group allocation probability `group_allocation_prob` as the mean of column `A` in data: 0.42
#> 
#> ── Symbolic differentiation of estimand function ──
#> 
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi0' as: '[1] -1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv0`
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi1' as: '[1] 1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv1`

print(ate)
#> Object of class 'rctglm'
#> 
#> Call:  rctglm(formula = Y ~ ., group_indicator = A, family = binomial, 
#>     data = dat_binom, estimand_fun = "ate")
#> 
#> - Counterfactual control mean (Psi_0=E[Y|X, A=0]) estimate: 0.586
#> - Counterfactual control mean (Psi_1=E[Y|X, A=1]) estimate: 0.883
#> - Estimand function r: psi1 - psi0
#> - Estimand (r(Psi_1, Psi_0)) estimate (SE): 0.298 (0.0685)
estimand(ate)
#> [1] 0.2976071
se_estimand(ate)
#> [1] 0.06849311
coef(ate)
#> (Intercept)          X1           A 
#>   0.5794888   1.3539024   2.2197052 
summary(ate)
#> 
#> Call:
#> (function (formula, family = gaussian, data, weights, subset, 
#>     na.action, start = NULL, etastart, mustart, offset, control = list(...), 
#>     model = TRUE, method = "glm.fit", x = FALSE, y = TRUE, singular.ok = TRUE, 
#>     contrasts = NULL, ...) 
#> {
#>     cal <- match.call()
#>     if (is.character(family)) 
#>         family <- get(family, mode = "function", envir = parent.frame())
#>     if (is.function(family)) 
#>         family <- family()
#>     if (is.null(family$family)) {
#>         print(family)
#>         stop("'family' not recognized")
#>     }
#>     if (missing(data)) 
#>         data <- environment(formula)
#>     mf <- match.call(expand.dots = FALSE)
#>     m <- match(c("formula", "data", "subset", "weights", "na.action", 
#>         "etastart", "mustart", "offset"), names(mf), 0L)
#>     mf <- mf[c(1L, m)]
#>     mf$drop.unused.levels <- TRUE
#>     mf[[1L]] <- quote(stats::model.frame)
#>     mf <- eval(mf, parent.frame())
#>     if (identical(method, "model.frame")) 
#>         return(mf)
#>     if (!is.character(method) && !is.function(method)) 
#>         stop("invalid 'method' argument")
#>     if (identical(method, "glm.fit")) 
#>         control <- do.call("glm.control", control)
#>     mt <- attr(mf, "terms")
#>     Y <- model.response(mf, "any")
#>     if (length(dim(Y)) == 1L) {
#>         nm <- rownames(Y)
#>         dim(Y) <- NULL
#>         if (!is.null(nm)) 
#>             names(Y) <- nm
#>     }
#>     X <- if (!is.empty.model(mt)) 
#>         model.matrix(mt, mf, contrasts)
#>     else matrix(, NROW(Y), 0L)
#>     weights <- as.vector(model.weights(mf))
#>     if (!is.null(weights) && !is.numeric(weights)) 
#>         stop("'weights' must be a numeric vector")
#>     if (!is.null(weights) && any(weights < 0)) 
#>         stop("negative weights not allowed")
#>     offset <- as.vector(model.offset(mf))
#>     if (!is.null(offset)) {
#>         if (length(offset) != NROW(Y)) 
#>             stop(gettextf("number of offsets is %d should equal %d (number of observations)", 
#>                 length(offset), NROW(Y)), domain = NA)
#>     }
#>     mustart <- model.extract(mf, "mustart")
#>     etastart <- model.extract(mf, "etastart")
#>     fit <- eval(call(if (is.function(method)) "method" else method, 
#>         x = X, y = Y, weights = weights, start = start, etastart = etastart, 
#>         mustart = mustart, offset = offset, family = family, 
#>         control = control, intercept = attr(mt, "intercept") > 
#>             0L, singular.ok = singular.ok))
#>     if (length(offset) && attr(mt, "intercept") > 0L) {
#>         fit2 <- eval(call(if (is.function(method)) "method" else method, 
#>             x = X[, "(Intercept)", drop = FALSE], y = Y, mustart = fit$fitted.values, 
#>             weights = weights, offset = offset, family = family, 
#>             control = control, intercept = TRUE))
#>         if (!fit2$converged) 
#>             warning("fitting to calculate the null deviance did not converge -- increase 'maxit'?")
#>         fit$null.deviance <- fit2$deviance
#>     }
#>     if (model) 
#>         fit$model <- mf
#>     fit$na.action <- attr(mf, "na.action")
#>     if (x) 
#>         fit$x <- X
#>     if (!y) 
#>         fit$y <- NULL
#>     structure(c(fit, list(call = cal, formula = formula, terms = mt, 
#>         data = data, offset = offset, control = control, method = method, 
#>         contrasts = attr(X, "contrasts"), xlevels = .getXlevels(mt, 
#>             mf))), class = c(fit$class, c("glm", "lm")))
#> })(formula = Y ~ ., family = function (link = "logit") 
#> {
#>     linktemp <- substitute(link)
#>     if (!is.character(linktemp)) 
#>         linktemp <- deparse(linktemp)
#>     okLinks <- c("logit", "probit", "cloglog", "cauchit", "log")
#>     family <- "binomial"
#>     if (linktemp %in% okLinks) 
#>         stats <- make.link(linktemp)
#>     else if (is.character(link)) {
#>         stats <- make.link(link)
#>         linktemp <- link
#>     }
#>     else {
#>         if (inherits(link, "link-glm")) {
#>             stats <- link
#>             if (!is.null(stats$name)) 
#>                 linktemp <- stats$name
#>         }
#>         else {
#>             stop(gettextf("link \"%s\" not available for %s family; available links are %s", 
#>                 linktemp, family, paste(sQuote(okLinks), collapse = ", ")), 
#>                 domain = NA)
#>         }
#>     }
#>     variance <- function(mu) mu * (1 - mu)
#>     validmu <- function(mu) all(is.finite(mu)) && all(mu > 0 & 
#>         mu < 1)
#>     dev.resids <- function(y, mu, wt) .Call(C_binomial_dev_resids, 
#>         y, mu, wt)
#>     aic <- function(y, n, mu, wt, dev) {
#>         m <- if (any(n > 1)) 
#>             n
#>         else wt
#>         -2 * sum(ifelse(m > 0, (wt/m), 0) * dbinom(round(m * 
#>             y), round(m), mu, log = TRUE))
#>     }
#>     simfun <- function(object, nsim) {
#>         ftd <- fitted(object)
#>         n <- length(ftd)
#>         ntot <- n * nsim
#>         wts <- object$prior.weights
#>         if (any(wts%%1 != 0)) 
#>             stop("cannot simulate from non-integer prior.weights")
#>         if (!is.null(m <- object$model)) {
#>             y <- model.response(m)
#>             if (is.factor(y)) {
#>                 yy <- factor(1 + rbinom(ntot, size = 1, prob = ftd), 
#>                   labels = levels(y))
#>                 split(yy, rep(seq_len(nsim), each = n))
#>             }
#>             else if (is.matrix(y) && ncol(y) == 2) {
#>                 yy <- vector("list", nsim)
#>                 for (i in seq_len(nsim)) {
#>                   Y <- rbinom(n, size = wts, prob = ftd)
#>                   YY <- cbind(Y, wts - Y)
#>                   colnames(YY) <- colnames(y)
#>                   yy[[i]] <- YY
#>                 }
#>                 yy
#>             }
#>             else rbinom(ntot, size = wts, prob = ftd)/wts
#>         }
#>         else rbinom(ntot, size = wts, prob = ftd)/wts
#>     }
#>     structure(list(family = family, link = linktemp, linkfun = stats$linkfun, 
#>         linkinv = stats$linkinv, variance = variance, dev.resids = dev.resids, 
#>         aic = aic, mu.eta = stats$mu.eta, initialize = binomInitialize(family), 
#>         validmu = validmu, valideta = stats$valideta, simulate = simfun, 
#>         dispersion = 1), class = "family")
#> }, data = structure(list(Y = c(1L, 1L, 1L, 1L, 1L, 1L, 0L, 1L, 
#> 0L, 1L, 1L, 1L, 0L, 0L, 0L, 1L, 0L, 1L, 1L, 0L, 1L, 1L, 1L, 0L, 
#> 0L, 1L, 1L, 1L, 1L, 1L, 0L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 
#> 0L, 1L, 0L, 1L, 0L, 1L, 1L, 0L, 0L, 1L, 1L, 1L, 1L, 1L, 1L, 0L, 
#> 1L, 0L, 1L, 0L, 1L, 1L, 0L, 1L, 1L, 1L, 0L, 1L, 0L, 1L, 1L, 1L, 
#> 1L, 1L, 1L, 1L, 1L, 1L, 0L, 1L, 1L, 1L, 1L, 1L, 0L, 1L, 0L, 0L, 
#> 0L, 1L, 1L, 1L, 1L, 1L, 0L, 1L, 0L, 1L, 1L, 0L), X1 = c(-0.416115003400462, 
#> 0.982149585796568, 0.242180726748904, -0.937926854142168, 1.50585850038705, 
#> -1.0658511740201, 0.162270526694493, 0.260470392136449, -1.48820199304168, 
#> 1.41518386908496, 0.562507513278876, 0.642055740996767, -0.890532644083346, 
#> -0.116537516996528, -0.941974751187683, 1.11582791811755, 0.470135133706391, 
#> 0.860612712312033, -0.0703966456448681, -0.613180213640174, -0.336714959893361, 
#> -0.215840183865469, 0.621132286645758, -1.28402651676709, -1.30009244118702, 
#> -0.376769461656321, 0.103748646404691, -0.703562272582238, 1.49741394047522, 
#> -0.302826877559252, -1.37683133463726, 0.883170174632431, 0.671420279304275, 
#> 1.23002237503278, 1.64546787858301, -0.182926038663912, -1.37925781493758, 
#> 1.2506006007554, -1.87623677486404, -0.368787568278473, -2.12929730823288, 
#> 0.337646026995419, -1.57457484751255, -0.196319414848903, 0.915060107809175, 
#> 1.07863653772935, 1.4528028539521, -1.6532810892254, -2.76186751764091, 
#> 0.264369163537262, 1.02457159426044, -0.659185547986991, -0.386304378522672, 
#> -0.242498949352553, 0.220593240071621, -1.48109003020768, 0.724691204470142, 
#> -2.37751470126173, -0.0654092094259788, -0.469704813112509, -0.280516288067282, 
#> 0.57483658535669, -0.205676846062025, 0.234880505242246, 0.110628320902131, 
#> 0.274202232513379, 1.57491963007029, 0.567767831624168, -0.05961224987651, 
#> -0.149587714594292, 0.266168788377474, 0.528403242689259, -0.579388093114727, 
#> 0.94219937374031, -0.706361360151394, -1.27540210033337, 0.733259916535147, 
#> -0.420207094399273, -0.255375085688641, -0.369238216504303, 1.21244832967657, 
#> -0.44536300595727, 1.13866752147802, 0.190523483027336, -1.36639361234535, 
#> -1.29810260115606, -2.13709831177351, -0.0962055250040968, 0.694509054682261, 
#> -0.474591230032695, 1.76313061184639, 2.02584606385851, 0.427618808607201, 
#> -0.56647934926807, -0.83383531146376, 0.398259154519669, -1.47453793146991, 
#> -0.395441200477897, -0.130702467875736, -1.99949394951033), A = c(0L, 
#> 0L, 1L, 0L, 0L, 1L, 0L, 0L, 0L, 1L, 0L, 0L, 0L, 1L, 0L, 1L, 0L, 
#> 1L, 0L, 0L, 0L, 1L, 1L, 0L, 1L, 1L, 1L, 1L, 1L, 0L, 0L, 0L, 0L, 
#> 0L, 0L, 1L, 1L, 1L, 1L, 1L, 0L, 1L, 0L, 1L, 0L, 1L, 0L, 0L, 0L, 
#> 1L, 0L, 1L, 0L, 0L, 0L, 0L, 1L, 1L, 1L, 0L, 1L, 1L, 0L, 1L, 1L, 
#> 0L, 0L, 0L, 0L, 1L, 0L, 0L, 0L, 0L, 1L, 0L, 1L, 1L, 0L, 1L, 0L, 
#> 0L, 1L, 0L, 0L, 0L, 0L, 0L, 0L, 1L, 0L, 0L, 1L, 0L, 1L, 1L, 0L, 
#> 1L, 1L, 0L)), class = "data.frame", row.names = c(NA, -100L)))
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)   0.5795     0.3323   1.744 0.081141 .  
#> X1            1.3539     0.3263   4.149 3.34e-05 ***
#> A             2.2197     0.6701   3.313 0.000925 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 120.430  on 99  degrees of freedom
#> Residual deviance:  80.586  on 97  degrees of freedom
#> AIC: 86.586
#> 
#> Number of Fisher Scoring iterations: 5
#>