Use prognostic covariate adjustment when fitting an rctglm

The procedure uses fit_best_learner to fit a prognostic model to historical data and uses the model to produce counterfactual predictions as a prognostic score that is then adjusted for as a covariate in the rctglm procedure.

Usage

rctglm_with_prognosticscore(
  formula,
  family,
  data,
  exposure_indicator,
  exposure_prob = NULL,
  estimand_fun = "ate",
  estimand_fun_deriv0 = NULL,
  estimand_fun_deriv1 = NULL,
  cv_variance = TRUE,
  cv_variance_folds = 10,
  ...,
  data_hist,
  prog_formula = NULL,
  cv_prog_folds = 5,
  learners = default_learners(),
  verbose = options::opt("verbose")
)

Arguments

formula

an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’ in the glm documentation.

family

a description of the error distribution and link function to be used in the model. For glm this can be a character string naming a family function, a family function or the result of a call to a family function. For glm.fit only the third option is supported. (See family for details of family functions.)

data

an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which glm is called.

exposure_indicator

(name of) the binary variable in data that identifies randomisation groups. The variable is required to be binary to make the "orientation" of the estimand_fun clear.

exposure_prob

a numeric with the probabiliy of being in "group 1" (rather than group 0) in groups defined by exposure_indicator. As a default, the ratio of 1's in data is used.

estimand_fun

a function with arguments psi1 and psi0 specifying the estimand. Alternative, specify "ate" or "rate_ratio" as a character to use one of the default estimand functions. See more details in the "Estimand" section of this documentation.

estimand_fun_deriv0

a function specifying the derivative of estimand_fun wrt. psi0. As a default the algorithm will use symbolic differentiation to automatically find the derivative from estimand_fun

estimand_fun_deriv1

a function specifying the derivative of estimand_fun wrt. psi1. As a default the algorithm will use symbolic differentiation to automatically find the derivative from estimand_fun

cv_variance

a logical determining whether to estimate the variance using cross-validation (see details of rctglm).

cv_variance_folds

a numeric with the number of folds to use for cross validation if cv_variance is TRUE.

...

Additional arguments passed to stats::glm()

data_hist

a data.frame with historical data on which to fit a prognostic model

prog_formula

a character or numeric with the formula for fitting the prognostic model on the historical data data_hist. Default models the response (assumed same as in formula) using all columns in the data_hist data

cv_prog_folds

a numeric with the number of cross-validation folds used when fitting and evaluating models

learners

a list of tidymodels

verbose

numeric verbosity level. Higher values means more information is printed in console. A value of 0 means nothing is printed to console during execution (Defaults to 2, overwritable using option 'PostCard.verbose' or environment variable 'R_POSTCARD_VERBOSE')

Value

rctglm_with_prognosticscore returns an object of class rctglm_prog, which inherits from rctglm.

Due to inheritance, methods described in the Value section of rctglm and rctglm_methods also work for extracting information from this object.

An rctglm_prog object is a list with the same components as an rctglm object, but with an additional list element of:

prognostic_info: List with information about the fitted prognostic model on historical data. It has components:

formula: The formula with symbolic description of how the response is modelled as function of covariates in the models

model_fit: A trained workflow - the result of fit_best_learner

learners: A list of learners used for the discrete super learner

cv_folds: The amount of folds used for cross validation

data: The historical data used for cross validation when fitting and testing models

Details

More details on prognostic models and scores being predictions of counterfactual means in control group.

Examples

# Generate some data
n <- 100
b0 <- 1
b1 <- 1.5
b2 <- 2
W1 <- runif(n, min = -2, max = 2)
exposure_prob <- .5

dat_treat <- glm_data(
  Y ~ b0+b1*abs(sin(W1))+b2*A,
  W1 = W1,
  A = rbinom (n, 1, exposure_prob)
)

dat_notreat <- glm_data(
  Y ~ b0+b1*abs(sin(W1)),
  W1 = W1
)

ate <- rctglm_with_prognosticscore(
  formula = Y ~ .,
  exposure_indicator = A,
  exposure_prob = exposure_prob,
  data = dat_treat,
  family = gaussian(),
  estimand_fun = "ate",
  data_hist = dat_notreat)
#> 
#> ── Fitting prognostic model ──
#> 
#> ℹ Created formula for fitting prognostic model as: Y ~ .
#> ℹ Fitting learners
#> • mod_mars
#> • mod_lm
#> • mod_gbt
#> i	No tuning parameters. `fit_resamples()` will be attempted
#> i 1 of 3 resampling: mod_mars
#> ✔ 1 of 3 resampling: mod_mars (113ms)
#> i	No tuning parameters. `fit_resamples()` will be attempted
#> i 2 of 3 resampling: mod_lm
#> ✔ 2 of 3 resampling: mod_lm (86ms)
#> i 3 of 3 tuning:     mod_gbt
#> ✔ 3 of 3 tuning:     mod_gbt (891ms)
#> ℹ Model with lowest RMSE: mod_mars
#> ℹ Investigate trained learners and fitted model in `prognostic_info` list element
#> 
#> ── Symbolic differentiation of estimand function ──
#> 
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi0' as: '-1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv0`
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi1' as: '1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv1`

# Pull information on estimand
estimand(ate)
#>   Estimate Std. Error
#> 1 2.345047  0.2124371