Find the best learner in terms of RMSE among specified learners using cross validation

Find the best learner in terms of RMSE among specified learners using cross validation

Usage

fit_best_learner(
  data,
  formula,
  cv_folds = 5,
  learners = default_learners(),
  verbose = options::opt("verbose")
)

Arguments

data

an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which glm is called.

formula

an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.

cv_folds

a numeric with the number of cross-validation folds used when fitting and evaluating models

learners

a list of tidymodels. Default uses a combination of MARS, linear regression and boosted trees

verbose

numeric verbosity level specifying how much information should be printed to the user (Defaults to 2, overwritable using option 'PostCard.verbose' or environment variable 'R_POSTCARD_VERBOSE')

Value

a trained workflow

Examples

# Generate some "historical" data of a control treatment
n <- 100
w1 <- runif(n, min = -2, max = 2)
x1 <- abs(sin(w1))
a <- rbinom (n, 1, .5)
b0 <- 1
b1 <- 1.5
b2 <- 2

truemean_notreat <- b0+b1*x1
y_notreat <- rnorm(n, mean = truemean_notreat, sd = 1)
dat_hist <- data.frame(Y = y_notreat, W = w1)

# Fit a learner to the data with default learners
fit <- fit_best_learner(Y ~ ., data = dat_hist)
#> ℹ Fitting learners
#> • mod_mars
#> • mod_lm
#> • mod_gbt
#> i	No tuning parameters. `fit_resamples()` will be attempted
#> i 1 of 3 resampling: mod_mars
#> ✔ 1 of 3 resampling: mod_mars (172ms)
#> i	No tuning parameters. `fit_resamples()` will be attempted
#> i 2 of 3 resampling: mod_lm
#> ✔ 2 of 3 resampling: mod_lm (89ms)
#> i 3 of 3 tuning:     mod_gbt
#> ✔ 3 of 3 tuning:     mod_gbt (923ms)
#> ℹ Model with lowest RMSE: mod_gbt