Methods for objects of class rctglm

Methods mostly to extract information from model fit and inference. See details for more information on each method.

Usage

estimand(object)

# S3 method for class 'rctglm'
estimand(object)

est(object)

# S3 method for class 'rctglm'
coef(object, ...)

# S3 method for class 'rctglm'
print(x, digits = max(3L, getOption("digits") - 3L), ...)

Arguments

object

an object of class rctglm

...

additional arguments passed to methods

x

an object of class rctglm

digits

a numeric with the number of digits to display when printing

Details

The function estimand (or short-hand version est) can be used to extract a data.frame with an estimated value and standard error of the estimand.

A method for the generic coef has been added for rctglm (i.e., coef.rctglm), which uses the method coef.glm to extract coefficient information from the underlying glm fit in the procedure.

Examples

# Generate some data to showcase example
n <- 100
exposure_prob <- .5
dat_binom <- glm_data(
  Y ~ 1+1.5*X1+2*A,
  X1 = rnorm(n),
  A = rbinom(n, 1, exposure_prob),
  family = binomial()
)

# Fit the model
ate <- rctglm(formula = Y ~ .,
              exposure_indicator = A,
              exposure_prob = exposure_prob,
              data = dat_binom,
              family = binomial,
              estimand_fun = "ate")
#> 
#> ── Symbolic differentiation of estimand function ──
#> 
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi0' as: '-1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv0`
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi1' as: '1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv1`

print(ate)
#> 
#> Object of class rctglm 
#> 
#> Call:  rctglm(formula = Y ~ ., exposure_indicator = A, exposure_prob = exposure_prob, 
#>     data = dat_binom, family = binomial, estimand_fun = "ate")
#> 
#> Counterfactual control mean (psi_0=E[Y|X, A=0]) estimate: 0.6762
#> Counterfactual control mean (psi_1=E[Y|X, A=1]) estimate: 0.9097
#> Estimand function r: psi1 - psi0
#> Estimand (r(psi_1, psi_0)) estimate (SE): 0.2335 (0.076)
estimand(ate)
#>    Estimate Std. Error
#> 1 0.2334771 0.07600023
coef(ate)
#> (Intercept)          X1           A 
#>   0.8528778   1.0275222   1.7464931