Methods for objects of class rctglm

Methods mostly to extract information from model fit and inference. See details for more information on each method.

Usage

# S3 method for class 'rctglm'
print(x, digits = max(3L, getOption("digits") - 3L), ...)

# S3 method for class 'rctglm'
coef(object, ...)

# S3 method for class 'rctglm'
summary(object, ...)

# S3 method for class 'summary.rctglm'
print(x, digits = max(3L, getOption("digits") - 3L), ...)

estimand(x)

# S3 method for class 'rctglm'
estimand(x)

se_estimand(x)

# S3 method for class 'rctglm'
se_estimand(x)

Arguments

x

an object of class rctglm

digits

a numeric with the number of digits to display when printing

...

additional arguments passed to methods

object

an object of class rctglm

Details

coef just use the corresponding glm methods on the glm fit contained within the rctglm object. Thus, this function is a shortcuts to running coef(x$glm).

summary summarises information related to the estimand as well as the underlying GLM fit.

estimand and se_estimand are methods for extracting the estimated estimand value as well as the standard error (SE) of the estimand. Extract the plug-in estimation of the estimand, found by using the estimand function on the predicted counterfactual means of each group. These functions are a shortcut to extract the list elements estimand and se_estimandof an rctglm class object.

Examples

# Generate some data to showcase example
n <- 100
dat_binom <- glm_data(
  1+1.5*X1+2*A,
  X1 = rnorm(n),
  A = rbinom(n, 1, .5),
  family = binomial()
)

# Fit the model
ate <- rctglm(formula = Y ~ .,
              group_indicator = A,
              data = dat_binom,
              family = binomial,
              estimand_fun = "ate")
#> ℹ Setting the group allocation probability `group_allocation_prob` as the mean of column `A` in data: 0.42
#> 
#> ── Symbolic differentiation of estimand function ──
#> 
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi0' as: '[1] -1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv0`
#> ℹ Symbolically deriving partial derivative of the function 'psi1 - psi0' with respect to 'psi1' as: '[1] 1'.
#> • Alternatively, specify the derivative through the argument
#> `estimand_fun_deriv1`

print(ate)
#> 
#> Object of class 'rctglm'
#> 
#> Call:  rctglm(formula = Y ~ ., group_indicator = A, family = binomial, 
#>     data = dat_binom, estimand_fun = "ate")
#> 
#> Counterfactual control mean (psi_0=E[Y|X, A=0]) estimate: 0.5857
#> Counterfactual control mean (psi_1=E[Y|X, A=1]) estimate: 0.8833
#> Estimand function r: psi1 - psi0
#> Estimand (r(psi_1, psi_0)) estimate (SE): 0.2976 (0.06849)
estimand(ate)
#> [1] 0.2976071
se_estimand(ate)
#> [1] 0.06849311
coef(ate)
#> (Intercept)          X1           A 
#>   0.5794888   1.3539024   2.2197052 
summary(ate)
#> 
#> Call:  rctglm(formula = Y ~ ., group_indicator = A, family = binomial, 
#>     data = dat_binom, estimand_fun = "ate")
#> 
#> 
#> ── Summary of estimand related statistics: ──
#> 
#> Counterfactual means, psi0 and psi1, based on groups in column A
#> Estimand function r: psi1 - psi0
#> Estimand (r(psi_1, psi_0)) estimate (SE): 0.2976 (0.06849)
#> 
#> ── Summary of glm fit: ──
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)   0.5795     0.3323   1.744 0.081141 .  
#> X1            1.3539     0.3263   4.149 3.34e-05 ***
#> A             2.2197     0.6701   3.313 0.000925 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 120.430  on 99  degrees of freedom
#> Residual deviance:  80.586  on 97  degrees of freedom
#> AIC: 86.586
#> 
#> Number of Fisher Scoring iterations: 5